+++ R/MetaForest.R
@@ -4,13 +4,13 @@
 #' meta-analytic data. MetaForest is a wrapper for \link[ranger]{ranger}
 #' (Wright & Ziegler, 2015). As input, MetaForest takes the study effect sizes
 #' and their variances (these can be computed, for example, using the
-#' \link[metafor]{metafor} package), as well as the moderators that are to be
-#' included in the model. By default, MetaForest uses random-effects weights,
-#' and estimates the between-studies variance using a restricted
-#' maximum-likelihood estimator. However, it may be beneficial to first conduct
-#' an unweighted MetaForest, and then use the estimated residual heterogeneity
-#' from this model as the estimate of \code{tau2} for a random-effects weighted
-#' MetaForest.
+#' \code{\link[metafor:rma.uni]{metafor}} package), as well as the moderators
+#' that are to be included in the model. By default, MetaForest uses
+#' random-effects weights, and estimates the between-studies variance using a
+#' restricted maximum-likelihood estimator. However, it may be beneficial to
+#' first conduct an unweighted MetaForest, and then use the estimated residual
+#' heterogeneity from this model as the estimate of \code{tau2} for a
+#' random-effects weighted MetaForest.
 #' @param formula Formula. Specify a formula for the MetaForest model, for
 #' example, \code{yi ~ .} to predict the outcome \code{yi} from all moderators
 #' in the data.
@@ -20,6 +20,11 @@
 #' @param vi Character. Specify the name of the column in the \code{data} that
 #' contains the variances of the effect sizes. This column will be removed from
 #' the data prior to analysis. Defaults to \code{"vi"}.
+#' @param study Character. Specify the name of the column in the \code{data}
+#' that contains the study id, if the data include multiple effect sizes per
+#' study. This column can be a vector of integers, or a factor. This column will
+#' be removed from the data prior to analysis. See \code{Details} for more
+#' information about analyzing dependent data.
 #' @param whichweights Character. Indicate what time of weights are required.
 #' A random-effects MetaForest is grown by specifying \code{whichweights =
 #' "random"}. A fixed-effects MetaForest is grown by specifying
@@ -31,8 +36,9 @@
 #' split. Defaults to the square root of the number moderators (rounded down).
 #' @param method Character. Specify the method by which to estimate the residual
 #' variance. Can be set to one of the following: "DL", "HE", "SJ", "ML", "REML",
-#' "EB", "HS", or "GENQ". Default is "REML". See the \link[metafor]{metafor} package
-#' for more information about these estimators.
+#' "EB", "HS", or "GENQ". Default is "REML". See the
+#' \code{\link[metafor:rma.uni]{metafor}} package for more information about
+#' these estimators.
 #' @param tau2 Numeric. Specify a predetermined value for the residual
 #' heterogeneity. Entering a value here supersedes the estimated tau2 value.
 #' Defaults to NULL.
@@ -86,46 +92,30 @@
 #' summary(mf.bd2004, digits = 4)
 #' #Examine variable importance plot
 #' VarImpPlot(mf.bd2004)
-MetaForest <- function(formula, data, vi = "vi", whichweights = "random",
+MetaForest <- function(formula, data, vi = "vi", study = NULL,
+                       whichweights = "random",
                        num.trees = 500, mtry = NULL, method = "REML",
                        tau2 = NULL, ...) {
-    args <- match.call()
     if(grepl("(\\*|:|-)", formula[3])){
       stop("MetaForest only accepts additive model formulae, because the underlying regression trees algorithm captures interactions as a sequence of consecutive splits on the interacting variables.")
     }
-    yi <- as.character(formula[2])
-    mods <- get_all_vars(formula, data)
-    if(vi %in% names(mods)) mods <- mods[-match(vi, names(mods))]
-    vi <- data[[vi]]
-    data <- mods
+    cl <- match.call()
+    args <- as.list(cl)[-1]
 
-    rma_before <- metafor::rma(yi = data[[yi]], vi = vi, method = method)
+    df <- get_all_vars(formula, data = data)
+    if(vi %in% names(df)) df <- df[-match(vi, names(df))]
+    args[["v"]] <- data[[vi]]
+    args[["formula"]] <- paste(formula[2], formula[3],sep = " ~ ")
 
-    if(is.null(tau2)) tau2 <- rma_before$tau2
-
-    if (whichweights == "unif") {
-        metaweights <- rep(1, nrow(data))
+    if(is.null(study)){
+      args[["df"]] <- df
+      output <- do.call(MF, args)
+    } else {
+      if(study %in% names(df)) df <- df[-match(study, names(df))]
+      args[["df"]] <- df
+      args[["id"]] <- data[[study]]
+      output <- do.call(MF_cluster, args)
     }
-    if (whichweights == "fixed") {
-        metaweights <- (1/vi)
-    }
-    if (whichweights == "random") {
-        metaweights <- 1/(vi + tau2)
-    }
-
-    metaweights <- (metaweights/sum(metaweights)) * nrow(data)
-
-    mf <- ranger::ranger(formula = formula, data = data, num.trees = num.trees,
-                 mtry = mtry, importance = "permutation", write.forest = TRUE,
-                 case.weights = metaweights, ...)
-    mf$call <- formula
-    predicted <- mf$predictions
-    observed <- data[[yi]]
-    residuals <- observed - predicted
-
-    rma_after <- metafor::rma(yi = residuals, vi = vi, method = method)
-
-    output <- list(forest = mf, rma_before = rma_before, rma_after = rma_after, call = args, data = data, vi = vi, weights = metaweights)
-    class(output) <- "MetaForest"
+    output$call <- cl
     output
 }