+++ R/metaforest-deprecated.r
@@ -1,0 +1,13 @@
+## metaforest-deprecated
+#' @title Deprecated functions in package \pkg{metaforest}.
+#' @description Metaforest is actively maintained, and I strive to improve its
+#' functionality and ease of use. This sometimes means that a function can, or
+#' must, be deprecated.
+#'
+#' The functions listed below are deprecated and will be defunct in
+#' the near future. When possible, alternative functions with similar
+#' functionality are also mentioned. Help pages for deprecated functions are
+#' available at \code{help("-deprecated")}.
+#' @name metaforest-deprecated
+#' @keywords internal
+NULL