+++ NEWS.md
@@ -1,3 +1,13 @@
+##### Version 0.1.3
+* ClusterMF is soft deprecated; it MetaForest has the same functionality as 
+  ClusterMF. You can simply replace any call to ClusterMF with a call to
+  MetaForest with the same arguments.
+* A clustered MetaForest analysis no longer automatically doubles the number of
+  trees estimated. Instead, it divides num.trees trees by two, rounding up to 
+  the nearest even number.
+* Generic S3 methods are now properly declared as such, instead of being 
+  exported with their own documentation.
+
 ##### Version 0.1.2
 * Rewrote WeightedScatter to jointly plot numeric and factor variables
 * Rewrote PartialDependence to be an S3 generic, with methods for metaforest and